pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.7",
]
